Update naming style
